Make generation time long instead of int

+++ app/src/main/java/Blockchain/BlockchainEntry.java
@@ -1,4 +1,4 @@
 package Blockchain;
 
-public record BlockchainEntry(Block block, int generationTime) {
+public record BlockchainEntry(Block block, long generationTime) {
 }
